refactor: improve signature of l10n-translate-pipe for better type inference
In applications using "strictTemplates" angular compiler option the signature of the translate pipe is problematic: Regardless of the passed key, the return type always includes null.
In reality however, unless `null` or an empty string is passed, the pipe always returns a string.

This changes the signature to align with the pipes actual behaviour.